moved get_max_eig and orthogonalize to __utility.py
